server/requirements.txt
flask==3.0.0
flask-cors==6.0.2
flask-socketio==5.6.1
python-dotenv==1.2.2
requests==2.33.1
vaderSentiment==3.3.2
pydantic>=2.0,<3.0
apscheduler==3.11.2
eventlet==0.41.0
gunicorn>=21.0,<24.0
yfinance==1.2.0
PyJWT==2.9.0
psycopg2-binary==2.9.9
transformers>=4.40.0,<5.0.0
torch>=2.2.0,<3.0.0
